# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend CMS API untuk SGS Law Firm.\n\nAdmin API menggunakan JWT Bearer Token.\nSebagian besar admin endpoint juga membutuhkan X-Timestamp, X-Nonce, dan X-Signature.\nUpload gambar hanya membutuhkan JWT karena menggunakan multipart/form-data.\n\nFormat canonical X-Signature:\nMETHOD + newline + REQUEST_URI + newline + TIMESTAMP + newline + NONCE + newline + SHA256_BODY",
        "title": "SGS CMS API",
        "contact": {
            "name": "SGS CMS Development Team"
        },
        "license": {
            "name": "Private"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/news": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin News"
                ],
                "summary": "Daftar news admin",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Featured filter",
                        "name": "is_featured",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerPaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.NewsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin News"
                ],
                "summary": "Membuat news",
                "parameters": [
                    {
                        "description": "News data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNewsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.NewsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/news/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin News"
                ],
                "summary": "Detail news admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.NewsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin News"
                ],
                "summary": "Memperbarui news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "News data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateNewsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.NewsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "description": "Melakukan soft delete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin News"
                ],
                "summary": "Menghapus news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateNewsRequest": {
            "type": "object",
            "required": [
                "content",
                "status",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 100000,
                    "minLength": 10
                },
                "excerpt": {
                    "type": "string",
                    "maxLength": 1000
                },
                "featured_image_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "is_featured": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
...
        "dto.NewsResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_image_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.SwaggerPaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Data retrieved successfully"
                },
                "meta": {
                    "$ref": "#/definitions/dto.SwaggerPaginationMeta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.SwaggerPaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 25
                },
                "total_pages": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.SwaggerSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.UpdateNewsRequest": {
            "type": "object",
            "required": [
                "content",
                "status",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 100000,
                    "minLength": 10
                },
                "excerpt": {
                    "type": "string",
                    "maxLength": 1000
                },
                "featured_image_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "is_featured": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "XNonce": {
            "description": "Nilai unik untuk mencegah replay attack.",
            "type": "apiKey",
            "name": "X-Nonce",
            "in": "header"
        },
        "XSignature": {
            "description": "HMAC-SHA256 signature dari canonical request.",
            "type": "apiKey",
            "name": "X-Signature",
            "in": "header"
        },
        "XTimestamp": {
            "description": "Unix timestamp dalam detik.",
            "type": "apiKey",
            "name": "X-Timestamp",
            "in": "header"
        }
    }
}